Move YAML related code into its own package
