+++ dashboard/dashboard.go
@@ -1,4 +1,4 @@
-package grabana
+package dashboard
 
 import (
 	"encoding/json"
@@ -20,40 +20,40 @@ type TagAnnotation struct {
 	Tags       []string
 }
 
-// DashboardBuilderOption represents an option that can be used to configure a
+// Option represents an option that can be used to configure a
 // dashboard.
-type DashboardBuilderOption func(dashboard *DashboardBuilder)
+type Option func(dashboard *Builder)
 
-// DashboardBuilder is the main builder used to configure dashboards.
-type DashboardBuilder struct {
+// Builder is the main builder used to configure dashboards.
+type Builder struct {
 	board *sdk.Board
 }
 
-// NewDashboardBuilder creates a new dashboard builder.
-func NewDashboardBuilder(title string, options ...DashboardBuilderOption) DashboardBuilder {
+// New creates a new dashboard builder.
+func New(title string, options ...Option) Builder {
 	board := sdk.NewBoard(title)
 	board.ID = 0
 	board.Timezone = ""
 
-	builder := &DashboardBuilder{board: board}
+	builder := &Builder{board: board}
 
-	for _, opt := range append(dashboardDefaults(), options...) {
+	for _, opt := range append(defaults(), options...) {
 		opt(builder)
 	}
 
 	return *builder
 }
 
-func dashboardDefaults() []DashboardBuilderOption {
-	return []DashboardBuilderOption{
+func defaults() []Option {
+	return []Option{
 		defaultTimePicker(),
 		defaultTime(),
 		SharedCrossHair(),
 	}
 }
 
-func defaultTime() DashboardBuilderOption {
-	return func(builder *DashboardBuilder) {
+func defaultTime() Option {
+	return func(builder *Builder) {
 		builder.board.Time = sdk.Time{
 			From: "now-3h",
 			To:   "now",
@@ -61,8 +61,8 @@ func defaultTime() DashboardBuilderOption {
 	}
 }
 
-func defaultTimePicker() DashboardBuilderOption {
-	return func(builder *DashboardBuilder) {
+func defaultTimePicker() Option {
+	return func(builder *Builder) {
 		builder.board.Timepicker = sdk.Timepicker{
 			RefreshIntervals: []string{"5s", "10s", "30s", "1m", "5m", "15m", "30m", "1h", "2h", "1d"},
 			TimeOptions:      []string{"5m", "15m", "1h", "6h", "12h", "24h", "2d", "7d", "30d"},
@@ -76,15 +76,20 @@ func defaultTimePicker() DashboardBuilderOption {
 // which your configuration management tool of choice can then feed into
 // Grafana's dashboard via its provisioning support.
 // See https://grafana.com/docs/grafana/latest/administration/provisioning/#dashboards
-func (builder *DashboardBuilder) MarshalJSON() ([]byte, error) {
+func (builder *Builder) MarshalJSON() ([]byte, error) {
 	return json.Marshal(builder.board)
+}
+
+// Internal.
+func (builder *Builder) Internal() *sdk.Board {
+	return builder.board
 }
 
 // VariableAsConst adds a templated variable, defined as a set of constant
 // values.
 // See https://grafana.com/docs/grafana/latest/reference/templating/#variable-types
-func VariableAsConst(name string, options ...constant.Option) DashboardBuilderOption {
-	return func(builder *DashboardBuilder) {
+func VariableAsConst(name string, options ...constant.Option) Option {
+	return func(builder *Builder) {
 		templatedVar := constant.New(name, options...)
 
 		builder.board.Templating.List = append(builder.board.Templating.List, templatedVar.Builder)
@@ -94,8 +99,8 @@ func VariableAsConst(name string, options ...constant.Option) DashboardBuilderOp
 // VariableAsCustom adds a templated variable, defined as a set of custom
 // values.
 // See https://grafana.com/docs/grafana/latest/reference/templating/#variable-types
-func VariableAsCustom(name string, options ...custom.Option) DashboardBuilderOption {
-	return func(builder *DashboardBuilder) {
+func VariableAsCustom(name string, options ...custom.Option) Option {
+	return func(builder *Builder) {
 		templatedVar := custom.New(name, options...)
 
 		builder.board.Templating.List = append(builder.board.Templating.List, templatedVar.Builder)
@@ -104,8 +109,8 @@ func VariableAsCustom(name string, options ...custom.Option) DashboardBuilderOpt
 
 // VariableAsInterval adds a templated variable, defined as an interval.
 // See https://grafana.com/docs/grafana/latest/reference/templating/#variable-types
-func VariableAsInterval(name string, options ...interval.Option) DashboardBuilderOption {
-	return func(builder *DashboardBuilder) {
+func VariableAsInterval(name string, options ...interval.Option) Option {
+	return func(builder *Builder) {
 		templatedVar := interval.New(name, options...)
 
 		builder.board.Templating.List = append(builder.board.Templating.List, templatedVar.Builder)
@@ -114,8 +119,8 @@ func VariableAsInterval(name string, options ...interval.Option) DashboardBuilde
 
 // VariableAsQuery adds a templated variable, defined as a query.
 // See https://grafana.com/docs/grafana/latest/reference/templating/#variable-types
-func VariableAsQuery(name string, options ...query.Option) DashboardBuilderOption {
-	return func(builder *DashboardBuilder) {
+func VariableAsQuery(name string, options ...query.Option) Option {
+	return func(builder *Builder) {
 		templatedVar := query.New(name, options...)
 
 		builder.board.Templating.List = append(builder.board.Templating.List, templatedVar.Builder)
@@ -123,15 +128,15 @@ func VariableAsQuery(name string, options ...query.Option) DashboardBuilderOptio
 }
 
 // Row adds a row to the dashboard.
-func Row(title string, options ...row.Option) DashboardBuilderOption {
-	return func(builder *DashboardBuilder) {
+func Row(title string, options ...row.Option) Option {
+	return func(builder *Builder) {
 		row.New(builder.board, title, options...)
 	}
 }
 
 // TagsAnnotation adds a new source of annotation for the dashboard.
-func TagsAnnotation(annotation TagAnnotation) DashboardBuilderOption {
-	return func(builder *DashboardBuilder) {
+func TagsAnnotation(annotation TagAnnotation) Option {
+	return func(builder *Builder) {
 		builder.board.Annotations.List = append(builder.board.Annotations.List, sdk.Annotation{
 			Name:       annotation.Name,
 			Datasource: &annotation.Datasource,
@@ -144,43 +149,43 @@ func TagsAnnotation(annotation TagAnnotation) DashboardBuilderOption {
 }
 
 // Editable marks the dashboard as editable.
-func Editable() DashboardBuilderOption {
-	return func(builder *DashboardBuilder) {
+func Editable() Option {
+	return func(builder *Builder) {
 		builder.board.Editable = true
 	}
 }
 
 // ReadOnly marks the dashboard as non-editable.
-func ReadOnly() DashboardBuilderOption {
-	return func(builder *DashboardBuilder) {
+func ReadOnly() Option {
+	return func(builder *Builder) {
 		builder.board.Editable = false
 	}
 }
 
 // SharedCrossHair configures the graph tooltip to be shared across panels.
-func SharedCrossHair() DashboardBuilderOption {
-	return func(builder *DashboardBuilder) {
+func SharedCrossHair() Option {
+	return func(builder *Builder) {
 		builder.board.SharedCrosshair = true
 	}
 }
 
 // DefaultTooltip configures the graph tooltip NOT to be shared across panels.
-func DefaultTooltip() DashboardBuilderOption {
-	return func(builder *DashboardBuilder) {
+func DefaultTooltip() Option {
+	return func(builder *Builder) {
 		builder.board.SharedCrosshair = false
 	}
 }
 
 // Tags adds the given set of tags to the dashboard.
-func Tags(tags []string) DashboardBuilderOption {
-	return func(builder *DashboardBuilder) {
+func Tags(tags []string) Option {
+	return func(builder *Builder) {
 		builder.board.Tags = tags
 	}
 }
 
 // AutoRefresh defines the auto-refresh interval for the dashboard.
-func AutoRefresh(interval string) DashboardBuilderOption {
-	return func(builder *DashboardBuilder) {
+func AutoRefresh(interval string) Option {
+	return func(builder *Builder) {
 		builder.board.Refresh = &sdk.BoolString{Flag: true, Value: interval}
 	}
 }
